refactor: merge habit parser families back together
